+++ website/src/components/docs/components/demo/radio-button-group-demo.tsx
@@ -9,9 +9,9 @@ import {
 export const RadioButtonGroupDemo = (props: RadioGroupProps) => {
   const options = [{ value: 'S' }, { value: 'M' }, { value: 'L', disabled: true }, { value: 'XL' }]
   return (
-    <RadioButtonGroup defaultValue="react" orientation="vertical" {...props}>
+    <RadioButtonGroup defaultValue="M" {...props}>
       {options.map((option, id) => (
-        <Radio key={id} value={option.value} disabled={option.disabled}>
+        <Radio key={id} value={option.value} disabled={option.disabled} px="0">
           <RadioControl />
           <RadioLabel>{option.value}</RadioLabel>
         </Radio>