+++ website/src/components/marketing/theme/color-palette-picker.tsx
@@ -1,6 +1,11 @@
-import { Circle, Grid, Stack } from 'styled-system/jsx'
+import { Circle, Stack } from 'styled-system/jsx'
 import { token } from 'styled-system/tokens'
-import { Button } from '~/components/ui/button'
+import {
+  Radio,
+  RadioButtonGroup,
+  RadioControl,
+  RadioLabel,
+} from '~/components/ui/radio-button-group'
 import { Typography } from '~/components/ui/typography'
 import { useColorMode } from '~/lib/use-color-mode'
 import { useThemeGenerator } from '~/lib/use-theme-generator'
@@ -14,37 +19,36 @@ export const ColorPalettePicker = () => {
       <Typography textStyle="sm" fontWeight="semibold">
         Color
       </Typography>
-      <Grid columns={3}>
+      <RadioButtonGroup
+        value={currentColorPalette.value}
+        size="sm"
+        variant="outline"
+        display="grid"
+        gridTemplateColumns="repeat(3, 1fr)"
+        onChange={(e) => {
+          updateColorPalette(
+            colorPalettes.find((colorPalette) => colorPalette.value === e.value) ??
+              currentColorPalette,
+          )
+        }}
+      >
         {colorPalettes.map((colorPalette, id) => (
-          <Button
-            key={id}
-            variant="secondary"
-            size="sm"
-            justifyContent="start"
-            onClick={() => updateColorPalette(colorPalette)}
-            {...(currentColorPalette.value === colorPalette.value
-              ? { 'data-selected': 'true' }
-              : {})}
-            _selected={{
-              background: 'initial',
-              color: 'fg.default',
-              borderColor: 'border.accent',
-              boxShadow: 'accent',
-            }}
-          >
-            <Circle
-              width="3.5"
-              height="3.5"
-              style={{
-                background: token.var(
-                  `colors.${colorPalette.value}.${colorMode === 'light' ? '600' : '300'}`,
-                ),
-              }}
-            />
-            {colorPalette.label}
-          </Button>
+          <Radio key={id} value={colorPalette.value} justifyContent="flex-start">
+            <RadioControl />
+            <RadioLabel>
+              <Circle
+                size="3.5"
+                style={{
+                  background: token.var(
+                    `colors.${colorPalette.value}.${colorMode === 'light' ? '600' : '300'}`,
+                  ),
+                }}
+              />
+              {colorPalette.label}
+            </RadioLabel>
+          </Radio>
         ))}
-      </Grid>
+      </RadioButtonGroup>
     </Stack>
   )
 }